requirements.txt
torch
torchvision
click
pyyaml
pillow
pycocotools
numpy
matplotlib
tqdm
gdown
scikit-learn
torch-geometric
catboost
opencv-python
transformers
timm
safetensors
albumentations
